BigInteger for Product model

+++ src/main/java/com/appcharge/server/models/player/Product.java
@@ -2,11 +2,12 @@ package com.appcharge.server.models.player;
 
 import lombok.Getter;
 import lombok.Setter;
+import java.math.BigInteger;
 
 @Getter
 @Setter
 public class Product {
-    private int amount;
+    private BigInteger amount;
     private ProductId productId;
     private String sku;
     private String name;
